Description of project
Gallerymage is a  web application that allows users to explore and search for high-quality images from various categories. The application fetches random images from the Unsplash API and provides an intuitive interface for users to navigate through the gallery. Users can also search for specific images using the search bar.